requirements.txt
pypdf2
fastapi
uvicorn
openai
python-dotenv
PyMuPDF
pandas
openpyxl
python-docx
tiktoken